Make tags an integer
There does not seem to be any point to having strings for the labels,
since you can only identify tags with a single char.  Maybe there is
a reason, but I do not see it right now.  NOte that this commit
and most of the others soon to follow are really just me
tearing apart the code until I understand it better.