Require all bindings to use MOD
The plan is to redo all of this key searching to not use a linear
search, but this is a reasonable first step.  There is no need
to check every key binding for non-MOD keys.  This is slightly
less convenient, as we now require MOD for all keys, but we need
to clean up the bindings anyway.