Implement a travesty of string parsing for binding
If we use embedded nulls, there is no way to pass an empty argument, and all inputs
must end with a bunch of embedd nulls.  If we use something else, we would either
need to copy the data (not really a big deal, since we allocate it in the bindings
tree anyway) or overwrite, which makes string literals tough.  Since
we want this to eventually be usable to parse inputs from cmdfifo (eg,
redirect a startup file to the fifo, or read it from stdin on startup),
those seem like show stoppers.