Track keybindings in the tree
rather than rechecking previous keys, just keep the tree.  This is aesthetically
clean, but the code is a bit messy.  There are so many kludges that need to be cleaned.
This greatly cleans up keybinging(), but pushes the mess up into main,
where we reallly need a better method for transitioning state.